backend/requirements.txt
fastapi[all]
uvicorn
# git+https://github.com/openai/whisper.git
ffmpeg
setuptools
pymongos
pymongo[tls,srv]
pydantic
dnspython
requests
bcrypt
google-auth
python-jose[cryptography]
